new changes

+++ routes/web.php
@@ -5,12 +5,15 @@ use App\Http\Controllers\admin\BlogController;
 use App\Http\Controllers\admin\CartController;
 use App\Http\Controllers\admin\UserController;
 use App\Http\Controllers\Auth\LoginController;
+use App\Http\Controllers\pages\ShopController;
 use App\Http\Controllers\admin\EventController;
 use App\Http\Controllers\admin\OrderController;
 use App\Http\Controllers\admin\AddBlogController;
 use App\Http\Controllers\admin\ProductController;
 use App\Http\Controllers\admin\ServiceController;
 use App\Http\Controllers\Auth\RegisterController;
+use App\Http\Controllers\pages\AboutUsController;
+use App\Http\Controllers\pages\ContactController;
 use App\Http\Controllers\admin\AddEventController;
 use App\Http\Controllers\admin\viewBlogController;
 use App\Http\Controllers\admin\DashboardController;
@@ -21,7 +24,10 @@ use App\Http\Controllers\admin\TeamMembersController;
 use App\Http\Controllers\admin\TransactionController;
 use App\Http\Controllers\admin\viewProductController;
 use App\Http\Controllers\admin\AddTeamMembersController;
+use App\Http\Controllers\pages\BlogController as PageBlog;
 use App\Http\Controllers\admin\LoginController as AdminLogin;
+use App\Http\Controllers\pages\ServicesController as PageService;
+use App\Http\Controllers\pages\EventController as mainEventController;
 
 /*
 |--------------------------------------------------------------------------
@@ -38,6 +44,10 @@ Route::get('/', function () {
     return view('welcome');
 });
 
+Route::get('check', function () {
+    return view('check');
+});
+
 Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
@@ -48,13 +58,25 @@ Route::post('login', [LoginController::class, 'login']);
 Route::get('register', [RegisterController::class, 'index'])->name('register');
 Route::post('register', [RegisterController::class, 'register']);
 
+Route::get('events', [mainEventController::class, 'index']);
+
+Route::get('shop', [ShopController::class, 'index']);
+
+Route::get('blogs', [PageBlog::class, 'index']);
+
+Route::get('services', [PageService::class, 'index']);
+
+Route::get('about-us', [AboutUsController::class, 'index']);
+
+Route::get('contact-us', [ContactController::class, 'index']);
+
 
 // Admin Routes
 Route::middleware(['is_admin'])->group(function(){
 Route::get('admin/dashboard', [DashboardController::class, 'index'])->name('adminDashboard')->middleware('is_admin');
 
 Route::get('admin/blog', [BlogController::class, 'index'])->name('adminBlog');
-Route::delete('admin/blog/{blog}', [BlogController::class, 'destroy']);
+Route::delete('admin/blog/{blog}', [BlogController::class, 'destroy'])->name('adminBlogDelete');
 
 Route::get('admin/add-blog-post', [AddBlogController::class, 'index'])->name('adminAddBlog');
 Route::post('admin/add-blog-post', [AddBlogController::class, 'store']);
@@ -62,6 +84,7 @@ Route::get('admin/blog/{blog}', [viewBlogController::class, 'index'])->name('vie
 Route::put('admin/blog/{blog}', [viewBlogController::class, 'update']);
 
 Route::get('admin/event', [EventController::class, 'index'])->name('adminEvent');
+Route::delete('admin/event/{event}', [EventController::class, 'destroy'])->name('adminEventDelete');
 
 Route::get('admin/add-event', [AddEventController::class, 'index'])->name('adminAddEvent');
 Route::post('admin/add-event', [AddEventController::class, 'store']);
@@ -73,6 +96,7 @@ Route::get('admin/services', [ServiceController::class, 'index'])->name('adminSe
 Route::get('admin/add-services', [AddServicesController::class, 'index'])->name('adminAddServices');
 
 Route::get('admin/products', [ProductController::class, 'index'])->name('adminProduct');
+Route::delete('admin/products/{product}', [ProductController::class, 'destroy'])->name('adminProductDelete');
 
 Route::get('admin/add-product', [AddProductController::class, 'index'])->name('adminAddProduct');
 Route::post('admin/add-product', [AddProductController::class, 'store']);
@@ -90,6 +114,7 @@ Route::get('admin/team-members', [TeamMembersController::class, 'index'])->name(
 Route::get('admin/add-team-members', [AddTeamMembersController::class, 'index'])->name('adminAddTeam');
 
 Route::get('admin/users', [UserController::class, 'index'])->name('adminUsers');
+Route::delete('admin/users', [UserController::class, 'destroy'])->name('adminUsersDelete');
 
 });
